+++ docs/rules/no-bind-in-components.md
@@ -1,7 +1,7 @@
 # Ensures connect functions are not being called within Components (no-connect-in-components)
 
-`connectObservable` and `connectFactoryObservable` should be used to create a
-shared hook from an observable. They shouldn't be used directly in React Components
+`bind` should be used to create a shared hook from an observable. They
+shouldn't be used directly in React Components
 
 ## Rule Details
 
@@ -9,7 +9,7 @@ Examples of **incorrect** code for this rule:
 
 ```js
 const Counter = () => {
-  const [useCounter] = connectObservable(count$);
+  const [useCounter] = bind(count$);
   const counter = useCounter();
 
   return <div>{counter}</div>;
@@ -19,7 +19,7 @@ const Counter = () => {
 Examples of **correct** code for this rule:
 
 ```js
-const [useCounter] = connectObservable(count$);
+const [useCounter] = bind(count$);
 const Counter = () => {
   const counter = useCounter();
