--- lib/index.js
@@ -1,23 +1,0 @@
-"use strict";
-
-//------------------------------------------------------------------------------
-// Requirements
-//------------------------------------------------------------------------------
-
-var requireIndex = require("requireindex");
-
-//------------------------------------------------------------------------------
-// Plugin Definition
-//------------------------------------------------------------------------------
-
-// import all rules in lib/rules
-module.exports.rules = requireIndex(__dirname + "/rules");
-
-module.exports.configs = {
-  recommended: {
-    plugins: ["react-rxjs"],
-    rules: {
-      "react-rxjs/no-connect-in-components": "error",
-    },
-  },
-};
