+++ rollup.config.js
@@ -1,0 +1,10 @@
+import typescript from '@rollup/plugin-typescript';
+
+export default {
+  input: 'src/index.ts',
+  output: {
+    file: 'dist/bundle.js',
+    format: 'cjs'
+  },
+  plugins: [typescript()]
+};
