+++ src/rules/index.ts
@@ -1,0 +1,5 @@
+import * as noBindInComponents from "./no-bind-in-components";
+
+export default {
+  "no-bind-in-components": noBindInComponents,
+};
